Index was float bugfix

+++ prepare_labels.py
@@ -113,12 +113,12 @@ def get_annotations(annotations_path: Path, working_dir: Path, frames: dict, res
     
     # Controlnet has difficulty generating people that are small - far away 
     if zoom_in: # TODO : crop_shape needs to be somewhere else too 
-        crop_shape = np.array([max_dif_x + min_pad , max_dif_y + min_pad])
+        crop_shape = np.array([max_dif_x + min_pad , max_dif_y + min_pad]).astype(int)
         # TODO : check that this updates the big dict
         for _, value in frame_dict.items():
             total_padding = crop_shape - (value["max"] - value["min"]) 
             value["candidates"][:,:2] = value["candidates"][:,:2] - value["min"] + total_padding/2
-            value["origin_x"], value["origin_y"] = value["min"] - total_padding/2
+            value["origin_x"], value["origin_y"] = (value["min"] - total_padding/2).astype(int)
             value["h"] = crop_shape[1]
             value["w"] = crop_shape[0]
